LayoutTests/media/loopend-limits-expected.txt

  
Test loopend attribute parsing, limits, and defaults.

<video loopend=1.5> - loopend specified, should not change
EXPECTED (video.loopEnd == '1.5') OK
EXPECTED (audio.loopEnd == '1.5') OK

<video end=1.0> - loopend not specified, should return end attr
EXPECTED (video.loopEnd == '1') OK
EXPECTED (audio.loopEnd == '1') OK

<video > - no loopend and no end, default should be infinity (end default)
EXPECTED (video.loopEnd == 'Infinity') OK
EXPECTED (audio.loopEnd == 'Infinity') OK

<video loopend='s1.0'> - bogus loopend and no end, default should be infinity (end default)
EXPECTED (video.loopEnd == 'Infinity') OK
EXPECTED (audio.loopEnd == 'Infinity') OK

<video loopend='s1.0' end=1.0> - bogus loopend, default should be end attr
EXPECTED (video.loopEnd == '1') OK
EXPECTED (audio.loopEnd == '1') OK
END OF TEST
